+++ bibfor/include/asterfort/PCHPDDMDumpAuxiliaryMat.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2022 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -26,7 +26,7 @@ interface
         IS :: is
         Mat :: aux
 #else
-        integer :: pc, is, aux
+        integer(kind=8) :: pc, is, aux
 #endif
     end subroutine PCHPDDMDumpAuxiliaryMat
 end interface
