+++ bibfor/include/asterfort/amumph.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -23,7 +23,7 @@
 interface
     subroutine amumph(action, solvez, matasz, rsolu, csolu, vcinez, nbsol, iret, prepos)
         character(len=*) :: action, matasz, vcinez, solvez
-        integer :: iret, nbsol
+        integer(kind=8) :: iret, nbsol
         real(kind=8) :: rsolu(*)
         complex(kind=8) :: csolu(*)
         aster_logical :: prepos
