+++ bibfor/include/asterfort/apetsc.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -25,8 +25,8 @@ interface
         character(len=*) :: matasz
         real(kind=8) :: rsolu(*)
         character(len=*) :: vcinez
-        integer :: nbsol
-        integer :: istop
-        integer :: iret
+        integer(kind=8) :: nbsol
+        integer(kind=8) :: istop
+        integer(kind=8) :: iret
     end subroutine apetsc
 end interface
