+++ bibfor/include/asterfort/apldlt.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,10 +20,10 @@
 !
 interface
     subroutine apldlt(kptsc, action, prepost, rsolu, vcine, nbsol)
-        integer, intent(in) :: kptsc
+        integer(kind=8), intent(in) :: kptsc
         character(len=*), intent(in) :: action, prepost
         real(kind=8), intent(inout) :: rsolu(*)
         character(len=19), intent(inout) :: vcine
-        integer, intent(in) :: nbsol
+        integer(kind=8), intent(in) :: nbsol
     end subroutine apldlt
 end interface
