+++ bibfor/include/asterfort/apvsmb.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@
 !
 interface
     subroutine apvsmb(kptsc, lmd, rsolu)
-        integer :: kptsc
+        integer(kind=8) :: kptsc
         aster_logical :: lmd
         real(kind=8) :: rsolu(*)
     end subroutine apvsmb
