+++ bibfor/include/asterfort/create_graph_comm.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@ interface
     subroutine create_graph_comm(object, type, nb_comm, comm, tag)
         character(len=*), intent(in) :: object
         character(len=*), intent(in) :: type
-        integer, intent(out) :: nb_comm
+        integer(kind=8), intent(out) :: nb_comm
         character(len=*), intent(in) :: comm, tag
     end subroutine create_graph_comm
 end interface
