+++ bibfor/include/asterfort/elg_calc_rhs_red.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@
 interface 
     subroutine elg_calc_rhs_red(matas1, nsecm, secm, solu2)
         character(len=19) :: matas1
-        integer :: nsecm
+        integer(kind=8) :: nsecm
         real(kind=8) :: secm(*)
         character(len=24) :: solu2
     end subroutine elg_calc_rhs_red
