+++ bibfor/include/asterfort/elg_calcxl_modal.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -25,7 +25,7 @@ interface
         PetscInt :: ke_mass
         real(kind=8) :: omega2
 # else
-        integer :: x1, vlag, ke_mass
+        integer(kind=8) :: x1, vlag, ke_mass
         real(kind=8) :: omega2
 # endif
     end subroutine elg_calcxl_modal
