+++ bibfor/include/asterfort/rsexch.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -24,8 +24,8 @@ interface
         character(len=*) :: kstop
         character(len=*) :: nomsd
         character(len=*) :: nomsy
-        integer :: iordr
+        integer(kind=8) :: iordr
         character(len=*) :: chextr
-        integer :: icode
+        integer(kind=8) :: icode
     end subroutine rsexch
 end interface
